[INLONG-4993][Manager] Refactor the unit tests class package
