--- inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/source/sqlserver/SqlServerSourceOperator.java
@@ -1,88 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.inlong.manager.service.source.sqlserver;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.inlong.manager.common.enums.ErrorCodeEnum;
-import org.apache.inlong.manager.common.enums.SourceType;
-import org.apache.inlong.manager.common.exceptions.BusinessException;
-import org.apache.inlong.manager.common.pojo.source.SourceRequest;
-import org.apache.inlong.manager.common.pojo.source.StreamSource;
-import org.apache.inlong.manager.common.pojo.source.sqlserver.SQLServerSource;
-import org.apache.inlong.manager.common.pojo.source.sqlserver.SQLServerSourceDTO;
-import org.apache.inlong.manager.common.pojo.source.sqlserver.SQLServerSourceRequest;
-import org.apache.inlong.manager.common.pojo.stream.StreamField;
-import org.apache.inlong.manager.common.util.CommonBeanUtils;
-import org.apache.inlong.manager.common.util.Preconditions;
-import org.apache.inlong.manager.dao.entity.StreamSourceEntity;
-import org.apache.inlong.manager.service.source.AbstractSourceOperator;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-/**
- * SQLServer source operator
- */
-@Service
-public class SQLServerSourceOperator extends AbstractSourceOperator {
-
-    @Autowired
-    private ObjectMapper objectMapper;
-
-    @Override
-    public Boolean accept(SourceType sourceType) {
-        return SourceType.SQLSERVER == sourceType;
-    }
-
-    @Override
-    protected String getSourceType() {
-        return SourceType.SQLSERVER.getType();
-    }
-
-    @Override
-    protected void setTargetEntity(SourceRequest request, StreamSourceEntity targetEntity) {
-        SQLServerSourceRequest sourceRequest = (SQLServerSourceRequest) request;
-        CommonBeanUtils.copyProperties(sourceRequest, targetEntity, true);
-        try {
-            SQLServerSourceDTO dto = SQLServerSourceDTO.getFromRequest(sourceRequest);
-            targetEntity.setExtParams(objectMapper.writeValueAsString(dto));
-        } catch (Exception e) {
-            throw new BusinessException(ErrorCodeEnum.SOURCE_INFO_INCORRECT.getMessage());
-        }
-    }
-
-    @Override
-    public StreamSource getFromEntity(StreamSourceEntity entity) {
-        SQLServerSource source = new SQLServerSource();
-        if (entity == null) {
-            return source;
-        }
-        String existType = entity.getSourceType();
-        Preconditions.checkTrue(getSourceType().equals(existType),
-                String.format(ErrorCodeEnum.SOURCE_TYPE_NOT_SAME.getMessage(), getSourceType(), existType));
-        SQLServerSourceDTO dto = SQLServerSourceDTO.getFromJson(entity.getExtParams());
-        CommonBeanUtils.copyProperties(entity, source, true);
-        CommonBeanUtils.copyProperties(dto, source, true);
-
-        List<StreamField> sourceFields = super.getSourceFields(entity.getId());
-        source.setFieldList(sourceFields);
-        return source;
-    }
-
-}
